warn

+++ Texture.h
@@ -29,6 +29,6 @@ public:
 	inline Vec2 getSize() { // use this vector direct to Prop2D::setScl(v)
 		int w,h;
 		getSize( &w, &h );
-		return Vec2(w,h);
+		return Vec2( (float)w, (float)h);
 	}
 };
